Introduce skeleton servant handlers
This also makes some other changes: the wallet is now split into an active
wallet and a passive wallet; this allows for cleaner initialization. We also
construct a specialized diffusion object for the wallet, which doesn't require
a monad parameter. This allows the wallet to be mostly isolated from everything
else.